Pascii
Pascii (pronounced Pass Key), uses a Private Key+User Name combo to
generate unique strong passwords, It also serves to replace a
traditional password vault applications (sort of), as it will always
generate the same password for a given user name, as long as the same
private key is used. The private key, is generated the first time you
run the application, the key can be automatically, or manually updated
via the configuration screen.